Handle killed windows.
The UnmapNotify handler now checks for window existance before updating
wm_state. This is a bit of a hack as I think we should be able to do
what we want but have X call our error handler if the window has died.

Instead, we'd see the X event loop freeze.