Store all root child windows, add stacklist.
RootWindow.children now has a Window for every child window (both
managed and unmanaged), the WM in now responsible for deciding which
ones to manage, references for which are now stored in stacklist.

Side-effect to this change is that now the root window gets the keypress
events. A couple of quick tests haven't told me why, but since I want
this behaviour when this code is moved to a separate xbindkeys-like
utility, I'll leave as is.