Ignore SIGCHLD instead of threads to avoid zombies.
Used to use threads to wait for a process to terminate and then read the
return code to avoid zombie process creation.  Popen.wait uses polling
to find out when a process ends which is a waste of resources.

Now, just ignore SIGCHLD, and drop all the threading code - much simpler!

Un*x provides.