EWMH clientlist and clientliststacking contain all.
Prior implementation of clientlist/stacking had only those windows
that belonged to the current desktop. This isn't correct so this
change puts all managed windows in those lists.

However, this only fixes half the problem. Need to refactor now so
that Client windows can be EWMH as well as the Root window...