Memoise only meaningful values.
Had a problem where the sizehints for emacs was being memoised while
it was still None. Not an ideal solution but it works for now.

Seems that emacs does not have all the window members set at
CreateNotify time (the point at which these are failing).

A future solution might be to instanciate the window at a later
point. eg, ConfigureRequest/MapRequest etc.